requirements.txt
# 📦 متطلبات Bot
# ================================

# Telegram Bot API
python-telegram-bot>=20.0

# تحميل الفيديوهات
yt-dlp>=2025.12.8

# اختياري: للتحميل الأسرع
aiohttp>=3.9.0
